Revert "Additions to remove resultsset"
This reverts commit a91d3071213de3cc388d7e69c99641f19ad57dee.
