+++ ipgm/Result.py
@@ -6,9 +6,12 @@ from ipgm.Candidacies import *
 
 #Single line of results
 class Result:
+	name = ''
 	results = {}
+	subset: list[Result] = []
 
-	def __init__(self, results: dict):
+	def __init__(self, name: str, results: dict):
+		self.name = name
 		self.results = results
 
 	@classmethod
@@ -23,6 +26,13 @@ class Result:
 	def createEmpty(self):
 		return Result('', {})
 
+	def contains(self, name: str) -> bool:
+		if self.name == name: return True
+		else:
+			for i in self.subset:
+				if i.contains(name): return True
+			return False
+	
 	def exportDict(self):
 		d = {}
 		d[self.name] = self.results
@@ -30,6 +40,17 @@ class Result:
 			d = appendDict(d, x.exportDict)
 		return d
 	
+	#Returns component given its name
+	def get(self, name: str) -> Result:
+
+		if self.name == name: return self
+		else:
+			for i in self.subset:
+				ig = i.get(name)
+				if ig != None: return ig
+			return None
+		#TODO: Check the auto-update (right now, a result isn't updated based on the results below)
+
 	def __repr__(self):
 		return '<{0}: {1}>'.format(self.name, self.results)
